Fix: Устранение ошибок сборки и выполнения, рефакторинг серверных дейс
Apply AI assistant changes to 2 files.
Ref: Fix: Устранение ошибок сборки и выполнения, рефакторинг серверных действий...
