+++ app/cyberfitness/types.ts
@@ -1,0 +1,36 @@
+import type { Database } from "@/types/database.types";
+
+type DbUser = Database["public"]["Tables"]["users"]["Row"];
+export type UserMetadata = DbUser['metadata'];
+
+export interface DailyActivityRecord {
+  date: string; 
+  filesExtracted: number;
+  tokensProcessed: number;
+  kworkRequestsSent: number; 
+  prsCreated: number;
+  branchesUpdated: number;
+  focusTimeMinutes?: number; 
+}
+
+export interface CyberFitnessProfile {
+  level: number; 
+  kiloVibes: number; 
+  focusTimeHours: number; 
+  skillsLeveled: number; 
+  activeQuests: string[]; 
+  completedQuests: string[]; 
+  unlockedPerks: string[]; 
+  achievements: string[]; 
+  cognitiveOSVersion: string; 
+  lastActivityTimestamp: string; 
+  dailyActivityLog: DailyActivityRecord[];
+  totalFilesExtracted: number; 
+  totalTokensProcessed: number; 
+  totalKworkRequestsSent: number; 
+  totalPrsCreated: number; 
+  totalBranchesUpdated: number; 
+  featuresUsed: Record<string, boolean | number | string>; 
+}
+
+export const CYBERFIT_METADATA_KEY = "cyberFitness";
